--- src/Routes/AppRoutes.jsx
@@ -1,5 +1,0 @@
-const AppRoutes = () => {
-  return <div></div>;
-};
-
-export default AppRoutes;
